implemented asyncronous user input (finally)
the raw_input function blocks the qt event loop, so it is not possible
to trigger background tasks with a QTimer and run a loop with raw_input
calls becasue the raw_input calls will keep the timers from emitting
their signals.

we now read user input in its own thread and communicate back and forth
with it using signals and slots.