BUGFIX: handle using control fan
The Stoker webpage displays a table with the sensors that are connected.
The table has the sensors organized by those being used to control a
fan, those used as alarms, and those used for monitoring. During
testing, the fan was not used, so the function that extracted sensor
information from the web page did not take this into account.

Sensors used for control, alarm, and monitoring are now handled
correctly.